refactor(provision): reorganize services
