+++ packages/ubuntu_provision/lib/keyboard.dart
@@ -1,4 +1,3 @@
 export 'src/keyboard/keyboard_l10n.dart';
 export 'src/keyboard/keyboard_model.dart' hide log;
 export 'src/keyboard/keyboard_page.dart';
-export 'src/keyboard/keyboard_service.dart';
