+++ packages/ubuntu_provision/lib/src/keyboard/keyboard_dialogs.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:ubuntu_localizations/ubuntu_localizations.dart';
+import 'package:ubuntu_provision/services.dart';
 import 'package:ubuntu_service/ubuntu_service.dart';
 import 'package:ubuntu_widgets/ubuntu_widgets.dart';
 import 'package:ubuntu_wizard/ubuntu_wizard.dart';
@@ -7,7 +8,6 @@ import 'package:yaru_widgets/yaru_widgets.dart';
 
 import 'keyboard_detector.dart';
 import 'keyboard_l10n.dart';
-import 'keyboard_service.dart';
 import 'keyboard_widgets.dart';
 
 const _kDialogWidthFactor = 0.65;
