+++ packages/ubuntu_provision/lib/src/timezone/timezone_model.dart
@@ -2,9 +2,8 @@ import 'package:collection/collection.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:timezone_map/timezone_map.dart';
 import 'package:ubuntu_logger/ubuntu_logger.dart';
+import 'package:ubuntu_provision/services.dart';
 import 'package:ubuntu_service/ubuntu_service.dart';
-
-import 'timezone_service.dart';
 
 final log = Logger('timezone');
