+++ packages/ubuntu_provision/lib/timezone.dart
@@ -1,4 +1,3 @@
 export 'src/timezone/timezone_l10n.dart';
 export 'src/timezone/timezone_model.dart';
 export 'src/timezone/timezone_page.dart';
-export 'src/timezone/timezone_service.dart';
