+++ packages/ubuntu_provision/lib/ubuntu_provision.dart
@@ -3,6 +3,7 @@ library ubuntu_provision;
 export 'flavor.dart';
 export 'keyboard.dart';
 export 'l10n.dart';
+export 'services.dart';
 export 'theme.dart';
 export 'timezone.dart';
 export 'widgets.dart';
