+++ packages/ubuntu_provision/test/keyboard/keyboard_detector_test.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:ubuntu_provision/src/keyboard/keyboard_detector.dart';
-import 'package:ubuntu_provision/src/keyboard/keyboard_service.dart';
+import 'package:ubuntu_provision/src/services/keyboard_service.dart';
 
 import 'test_keyboard.dart';
