+++ packages/ubuntu_provision/test/keyboard/keyboard_dialogs_test.dart
@@ -2,8 +2,8 @@ import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:ubuntu_provision/src/keyboard/keyboard_dialogs.dart';
-import 'package:ubuntu_provision/src/keyboard/keyboard_service.dart';
 import 'package:ubuntu_provision/src/keyboard/keyboard_widgets.dart';
+import 'package:ubuntu_provision/src/services/keyboard_service.dart';
 import 'package:ubuntu_service/ubuntu_service.dart';
 import 'package:ubuntu_test/ubuntu_test.dart';
