+++ packages/ubuntu_provision/test/keyboard/keyboard_model_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:platform/platform.dart';
 import 'package:ubuntu_provision/src/keyboard/keyboard_model.dart';
-import 'package:ubuntu_provision/src/keyboard/keyboard_service.dart';
+import 'package:ubuntu_provision/src/services/keyboard_service.dart';
 
 import 'test_keyboard.dart';
