+++ packages/ubuntu_provision/test/keyboard/keyboard_page_test.dart
@@ -6,8 +6,8 @@ import 'package:mockito/mockito.dart';
 import 'package:ubuntu_provision/src/keyboard/keyboard_l10n.dart';
 import 'package:ubuntu_provision/src/keyboard/keyboard_model.dart';
 import 'package:ubuntu_provision/src/keyboard/keyboard_page.dart';
-import 'package:ubuntu_provision/src/keyboard/keyboard_service.dart';
 import 'package:ubuntu_provision/src/keyboard/keyboard_widgets.dart';
+import 'package:ubuntu_provision/src/services/keyboard_service.dart';
 import 'package:ubuntu_service/ubuntu_service.dart';
 import 'package:ubuntu_test/ubuntu_test.dart';
 import 'package:ubuntu_widgets/ubuntu_widgets.dart';
