+++ packages/ubuntu_provision/test/keyboard/test_keyboard.dart
@@ -1,6 +1,7 @@
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:ubuntu_provision/keyboard.dart';
+import 'package:ubuntu_provision/services.dart';
 
 import 'test_keyboard.mocks.dart';
 export '../test_utils.dart';
