+++ packages/ubuntu_provision/test/keyboard/test_keyboard.mocks.dart
@@ -7,8 +7,9 @@ import 'dart:async' as _i4;
 import 'dart:ui' as _i5;
 
 import 'package:mockito/mockito.dart' as _i1;
+import 'package:subiquity_client/subiquity_client.dart' as _i2;
 import 'package:ubuntu_provision/src/keyboard/keyboard_model.dart' as _i3;
-import 'package:ubuntu_provision/src/keyboard/keyboard_service.dart' as _i2;
+import 'package:ubuntu_provision/src/services/keyboard_service.dart' as _i6;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
@@ -218,7 +219,7 @@ class MockKeyboardModel extends _i1.Mock implements _i3.KeyboardModel {
 /// A class which mocks [KeyboardService].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockKeyboardService extends _i1.Mock implements _i2.KeyboardService {
+class MockKeyboardService extends _i1.Mock implements _i6.KeyboardService {
   MockKeyboardService() {
     _i1.throwOnMissingStub(this);
   }
